Refactor of ResourceFile location and name
