+++ src/MooVC.Syntax.Tests/Concepts/ProjectTests/WhenConstructorIsCalled.cs
@@ -1,6 +1,7 @@
 namespace MooVC.Syntax.Concepts.ProjectTests;
 
 using MooVC.Syntax.Attributes.Project;
+using Resource = MooVC.Syntax.Attributes.Resource.Resource;
 
 public sealed class WhenConstructorIsCalled
 {
@@ -27,7 +28,7 @@ public sealed class WhenConstructorIsCalled
         Import import = ProjectTestsData.CreateImport();
         ItemGroup itemGroup = ProjectTestsData.CreateItemGroup();
         PropertyGroup propertyGroup = ProjectTestsData.CreatePropertyGroup();
-        Project.ResourceFile resourceFile = ProjectTestsData.CreateResourceFile();
+        Resource resource = ProjectTestsData.CreateResource();
         Sdk sdk = ProjectTestsData.CreateSdk();
         Target target = ProjectTestsData.CreateTarget();
 
@@ -37,7 +38,7 @@ public sealed class WhenConstructorIsCalled
             Imports = [import],
             ItemGroups = [itemGroup],
             PropertyGroups = [propertyGroup],
-            Resources = [resourceFile],
+            Resources = [resource],
             Sdks = [sdk],
             Targets = [target],
         };
@@ -46,7 +47,7 @@ public sealed class WhenConstructorIsCalled
         subject.Imports.ShouldBe(new[] { import });
         subject.ItemGroups.ShouldBe(new[] { itemGroup });
         subject.PropertyGroups.ShouldBe(new[] { propertyGroup });
-        subject.Resources.ShouldBe(new[] { resourceFile });
+        subject.Resources.ShouldBe(new[] { resource });
         subject.Sdks.ShouldBe(new[] { sdk });
         subject.Targets.ShouldBe(new[] { target });
         subject.IsUndefined.ShouldBeFalse();
