+++ src/MooVC.Syntax.Tests/Concepts/ResourceTests/WhenToDocumentIsCalled.cs
@@ -2,6 +2,7 @@ namespace MooVC.Syntax.Concepts.ResourceTests;
 
 using System.Xml.Linq;
 using MooVC.Syntax.Attributes.Resource;
+using Resource = MooVC.Syntax.Concepts.Resource;
 
 public sealed class WhenToDocumentIsCalled
 {
