+++ src/MooVC.Syntax.Tests/Concepts/ResourceTests/WhenValidateIsCalled.cs
@@ -3,6 +3,7 @@ namespace MooVC.Syntax.Concepts.ResourceTests;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using MooVC.Syntax.Attributes.Resource;
+using Resource = MooVC.Syntax.Concepts.Resource;
 
 public sealed class WhenValidateIsCalled
 {
