+++ src/MooVC.Syntax/Attributes/Resource/Resource.Scope.cs
@@ -1,0 +1,11 @@
+﻿namespace MooVC.Syntax.Attributes.Resource
+{
+    public partial class Resource
+    {
+        public enum Scope
+        {
+            Internal,
+            Public,
+        }
+    }
+}
