--- src/MooVC.Syntax/Concepts/Project.ResourceFile.Scope.cs
@@ -1,14 +1,0 @@
-﻿namespace MooVC.Syntax.Concepts
-{
-    public partial class Project
-    {
-        public partial class ResourceFile
-        {
-            public enum Scope
-            {
-                Internal,
-                Public,
-            }
-        }
-    }
-}
